Fix one instance of a crash when presented with bad DB state.
Plug one hole in the dike.  Don't blindly atoi() a DB return value.
One down, a zillion to go.  Fixed a couple of nits while in there.

Added TBCOREDIR define for the place where tmcd can go to die.
It is the directory where tmcd should run and is intended to insure that
it can always drop core.