Add a new timeout action; STATE
Allows the state_timeouts table to contain a new type of action
to take on timeout: STATE:newstate . This will force stated to
transition the node to newstate, and take any trigger actions
associated with that state.

We will use this to make timeouts in the secure reload path
force the node into the SECVIOLATION state.

Not yet tested.