Disable required client authentication on boss
We need to figure out what we are going to do about this because we
don't want to put client certs on all the dongles.  Besides, node id
is already present with the identity being encrypted by the SRK etc.

But how can boss differentiate between booting nodes with no client keys
and nodes requesting info normally from tmcd?