Ignore URIs that are not publicids in SubjAltName.
If a certificate has more than one URI in the Subject Alt Name
extension field the last URI was chosen as the URN for the
certificate. This choice was made regardless of the prefix of the
URN. The regex was modified to look for URIs that begin with
"urn:publicid:" to increase the chance that the right one is chosen.

If a certificate has more than one URI that begins with
"urn:publicid:", the last one is chosen. This is an improvement over
the previous state wherein the last URI was chosen regardless of prefix.