Roll my own ASCII to binary conversion and put quote size back to 1024
sscanf() with "%2x" requires an unsigned int * and writes sizeof(int) to
the pointer.  I made my own conversion because we were writing NULLs to
neighbor strings since we were passing characters to sscanf() and the
remaining three bytes were 00.  I feel there must be a library function
to do this two character ASCII to byte conversion though.

I forgot that the quote would be written in ASCII so I have to double
quote's size.