Avoid dynamic use (unpacking) of Python M2Crypto .egg file.
The FreeBSD port just installs the zipped .egg file for M2Crypto and there
is no option to make or install an unzipped version.  Unfortunately, the
zipped version gets unpacked, per-user when it is first used and the
directory it is unpacked to is not always writable.  To avoid unsightly
"permission denied" failures, we run a post-package-install phase to
manually unpack the .egg in the python site-library directory.  This is
wicked ugly and I really should be flogged for doing this...