Make double-sure that we don't do install from these directories in a devtree.
They are not in the top level GNUmakefile install, but if someone were to
manually install from their devtree, they would clobber the "master" versions.
This is because of the way that INSTALL_RCDIR and INSTALL_APACHE_CONFIG are
set in Makeconf.in.  They are not prefixed with ${prefix} since they can be
set directly via configure command line "--with" arguments.