Fix an argument parsing error.
In our sed 's///' commands applied to loadinfo we were assuming that if a
variable pattern didn't match, we would get back an empty string.  But what
we really get back is the entire loadinfo string.  Now sure how this ever
worked (if it ever did!)

Also did the piss-on-it thing.