Do not automatically merge updates to passwd/group/shadow files.
Now, if experimenters add users or groups to their nodes, and take
an image, they must explicitly tell prepare to overwrite the master
passwd/group/shadow files in /etc/emulab.  For Linux,
os_account_cleanup tried to update the masters automatically, but
this is not fullproof and can result in the masters getting passwd
hashes from shadow.  Now, a diff of the non-shadow files is printed
from prepare so that at least people can see that the masters need
to be manually updated.