Improved handling of trivial merges in gitmail
Use 'git diff-tree --no-commit-id --cc' on merge commits to determine if
there are any differences between the merge commit and its parents.
This is the case if any new files were hand-modified or added as part of
the merge (such as when conflicts were resolved).  If git-diff-tree
produces no output, the merge was trivial.

If the merge was trivial, the commit message is checked for an empty
body.  If empty, the commit is skipped when generating the email.  This
way if the committer documents the merge in some way (list of changes,
reason for the merge, etc.) the email includes that information.