Run frisbeelauncher as admin for subbosses
Subbosses make XMLRPC requests as the user 'elabman' which doesn't
belong to any projects, and therefore has no rights to non-global
images.  The easiset way to solve this is to run frisbeelauncher with
admin privileges when a subboss requests it.

We already check to make sure the requesting user has permission to use
the image in libosload well before we hand the request to the subboss,
so if the subboss requests the image we can trust it.