Better error message when requesting nodes of a nonexistent type
If a request was made for a node type that was not present in the physical
topology, then the error message was, "Too much bandwidth on emulated links."
This change adds a flag in mapping_precheck() that remembers whether the
requested node type was found.  If not, then a new error message is printed
instead of the bandwidth one.