Update NS version used by nsverify from 2.29 to 2.34.
This allows nsverify to be built on FreeBSD 7+.  Based on patches from John
Hickey <jhickey@isi.edu> and Serge van Ginderachter <serge@vanginderachter.be>,
thanks guys!

I also moved all the logic for downloading/patching the tarball from the
GNUmakefile into ns-patch.sh to bring this in line with our other third-party
packages that do the same.  This is in pursuit of one day having a GUDS
(Grand Unified Download Script).