Fix for some really stupid Cisco behavior
Some Ciscos limit the number of trunks on which you can configure
'extended range' VLANs (above 1k). This includes clearing them from
the trunk (eg. you don't actually have to be *using* them to get
this error). So, as a workaround, if the stack is not conifgured
to use VLANs over 1k, don't clear this range when cleaning out
trunk ports.