Resurrect disabling of UUID checks for AM.
The protogeni-wrapper change inadvertently lost the disabling of UUID
checks in the AM. The AM module is now recognized and certs without
UUIDs are allowed to pass through rather than generating an error.

This change required that the path parsing be moved earlier. The path
parse itself remains exactly the same, it just happens earlier on when
processing a request.