Restore regex mods that were lost.
In the change to protogeni-wrapper the URN regexes for the AM API were
lost. The existing regexes would pick the wrong URN out of PlanetLab
certificates, getting the second urn which is a UUID (urn:uuid:)
instead of the first (urn:publicid:IDN:). The regexes have been
modified to look for "urn:publicid:" instead of just "urn:".