Linux MFS: add support for cached PXE DHCP lease
Recently I modified Grub's linux initrd loader to append a cpio
archive containing the cached PXE DHCP ACK packet to the loaded
initrd/initramfs.  The Linux MFS now looks for this file when
configuring the network, and will use it to configure the interface
instead of sending out DHCP DISCOVER requests.  If the file is not
found then DHCP will be used to configure the control net interface.