Fix fixed interfaces for the fourth time
Okay, I know I said this the first three times, but I think I've got it
this time - the problem had to do with how I was handling the flipped
flag (of course). Many of the calls find_best_link() give the dest_pv
of a link instead of the pv (source pnode). When we do this, we have to
invert the flipped flag, 'cause we're effectively reversing the order
of the link.