Cleanup of rc.ipod (Linux version)
The awk implementation used in Busybox (for the Linux MFS) was giving
strange results when converting our dotted-quad IP address and mask to
a base-10 decimal integer.  It looks like it has something to do with
the sign bit in one of the intermediate values, and I don't really
want to debug it right now.

The kernel is perfectly happy with hex numbers for integer values when
writing to /proc/sys/* (directly or via sysctl), so we now simply pass
the dotted-quad address to awk (with '.' as the field separator) and
do a simple printf to get the result we want.

We no longer need to call sed either, since POSIX "remove matching
prefix/suffix" syntax is supported by ash in the MFS.