Handle failed XML-RPC calls
If TestBusy() or ClearAddress() fails due to connection problems, the
subboss_images table will be left in a bad state.  tmcd will assume
that the frisbeed process is still running on the subboss and will
return that address to nodes calling loadinfo.  These nodes will get
stuck waiting for a frisbeed that isn't running.

To prevent this, we now restart frisbeed when the TestBusy() or
ClearAddress() calls fail.  Hopefully, the next time around the calls
will succeed and subboss_frisbeelauncher can reset the DB state and
exit normally.