This reverts commit 0450a43c9924747ad5b9d1db0552a040be2b75ba.
This patch added a bunch of NULL checks throughout TMCD.  While most of
the NULL checks did handle it properly some didn't, so I'd rather leave
it all be and let it be fixed properly later.

Such a patch shouldn't be snuck in with all the TDLS changes, anyway.

Conflicts:

	tmcd/tmcd.c