Deal with non-numeric OIDs.
If a counter returns a non-numeric value, then treat it as zero and print
a warning.  Most likely these are OIDs that are unimplemented.
It will only warn once per device per OID, but even that might get tedious
so warnings can be turned off entirely with a global variable.