Fifth attempt at fixing fixed interfaces in assign
Think I got it this time (yeah, I've said that four times before).
It requires some extra bookkeeping; we were recording source and
dest interface names in tb_plink, but not source and dest node
names. With this new information, find_best_link() now has sole
responsibility for figuring out if the vlink and plink go in the
same direction, and the callers of find_best_link() are now
responsible for knowing which direction they are traversing the
vlink in (and thus whether they should tell find_best_link() to
check the source or dest of the vlink.)