Add a script to compress old expinfo directories.
Run it as "archive-expinfo [-t threshold]", where "threshold" is the
number of days experiments must have been inactive to be compressed
(defaulting to 1000).

Directories will be tarred and compressed in place.  For example, if
/usr/testbed/expinfo/testbed/example/1234 has been inactive for longer
than the threshold, its entire contents will be archived in
/usr/testbed/expinfo/testbed/example/1234.tar.bz2 and the original
directory removed.

The compression is extremely efficient, typically reducing directories
to around 2% of their original size.