Add autoconf variable to control use of SelfLoader.
This is a fer-now hack.  Perl 5.10 has issues with bogus taint check
triggers that appear quite often when using the SelfLoader.  Now if you put:
    SELFLOADER_DATA=""
In your defs-* file, it will disable the uses that cause problems.  Yes,
the configure script should figure out if this is needed for you, but...
another time.