Okay, that last commit actually added ssh transfer support to creating images.
The commit got away from me.  I was just going to check in a bug fix but got
a lot more.  The "-s" option of create_image will use ssh to collect an image
rather than having the node copy it across NFS.  This is part of the grand
unNFS-ification option which will get checked in tomorrow.  But this part
should stand alone as of this commit.