Do not register slices at the CH immediately; let the sa_daemon do that offline a couple of minutes later. Reduces the load on the CH when doing parallel registratons.
Add caching of user and slice credentials, to avoid regenerating the
same credentials over and over for the test scripts. Stored in the
geni_credentials table, they are checked for expiration before handing
them out. Also check to make sure that the user certificate has not
changed, and regen/cache if it has.