Make it work with NOSHAREDFS.
If LVM is not being used, then the code will attempt to make the extra FS
and use that for /vz.  However, in the NOSHAREDFS world, that filesystem
has already been made and is mounted on /local.  So check for that case and
just create a subdir/symlink: /vz -> /local/vz.