Enable new event scheduler
Use the new event scheduler by default for any
users/groups/experiments configured for it using EmulabFeatures.

Added '-N' option to eventsys.proxy to have it use the new scheduler
(if it exists) instead of the old one.

Added test for NewEventScheduler feature to eventsys_control so that
it will pass the '-N' option to eventsys.proxy for
experiments/users/groups using the feature.

These changes will go away when the new event scheduler replaces the
old one.