Allow tiptunnel to set serial device options via capture.
For now, I made it possible to set speed of the capture with
tiptunnel.  Later we could expose more options... but the right
thing to do is build a ConsoleAgent that is much more flexible.
I think we now know what we need as far as a good feature set
for that program.