If we don't flash the switch, restore its osid/partition info.
If a switch reload fails, and we're sure we didn't flash the
switch, then revert the changes to osid info in the nodes table
and the partitions table entries.  This way, we don't lie about
what's actually on the switch.  Otherwise, there could be a
problem on a subsequent reload -- a user could request something
that failed to load last time, but Emulab thinks that's what's
really on the switch, so it doesn't reload... but that wasn't
really what was installed on the switch.  So, this will help
us stay in synch.

Also remove all the os_select script invocations in favor of
using the Node->OSSelect method instead.  Save a fork, save a
life.