Fix a bug in the reorderLinks() function
Because it's operating on char*s, the STL list remove() function
doesn't work; it needs to do a strcmp, not a simple comparison. So, hand-write
the loop to do this.

Also add a santity check to produce a fatal error rather than allowing this
function to go into an infinite loop.