Fix potential wild pointer bug in reorderLinks()
Was returning a structure with a pointer to a stack-allocated string in
it. Use knowledge about where that string came from to avoid having to
make a copy of it, instead.

There is still a minor memory leak, but left behind a note rather than
fixing it (which would be quite involved).