Re-think no_connect violations
Re-think a very old idea in assign - that no_connect violations for
unmapped links should not be assessed against links whose endpoints are
not mapped. The problem with this is that assign is too hesitant to map
nodes - if there is more than one link that can't be mapped or whose
other end isn't mapped yet, then assign thinks it's better to leave the
node unmapped. This leads to confusing solutions (in which nodes remain
unmapped when really it's the links that are the problem), and I beleive
it keeps assign from exploring valuable parts of the solution space.

This patch chanes the behavior, so that no_connect violations get scored
even when the endpoints of the link are not mapped.

Not well enough tested to run in production.