Bugfix, and protect switch flashes via feature.
We now protect switch reloads via the SwitchFlash feature.  We
don't really give people a lot of advance warning about this,
and we wait long enough that it will irritatingly fail your
experiment swapin.  Ideally we should check before anything
starts happening, but we don't know if we *have* to reload
until we see what's on the nswitch node that gets allocated
to us... and by the time nodes are allocated, I think they
end up having to go through reloading if *any* of them fails.
Not sure though.

Also fixup handling of multiple libossetup_switch ops; if one
fails, abort the list instead of continuing on.