Inching toward recursive use, ala what is needed for elabinelab or subbosses.
Add the ability of the master server to have a "parent" from which it can
download an image if it doesn't have it or if the image is out of date.
Had to add some more goo to the GET reply, notably a hash so that we can
check for up-to-dateness.

The actual part where we upcall to the parent isn't done yet, that is why
this is "inching toward" and not "leaping and bounding toward"...

Also redid the child process management to not use SIGCHLD, no need for that.