Hierarchical downloading is now implemented.
The master server can be started with a "parent" mserver that it can call
if it doesn't have an image.  The master server will return an EAGAIN type
error to any clients that contact it while it is downloading an image from
its parent.  The client now has a "-B N" option to tell it to try again
every N seconds as long as it gets back that error.  This is the "store and
forward" mode.  The mserver also has a "cut through" style where it will
return to clients the mcast info it got from its parent so they can download
directly from the parent until the local mserver has it.