Make unicast transfers work.
There are some hacks involved here right now since a unicast frisbee server
currently can only support a single client.  So for now, we only allow a
single unicast server for any image (i.e., because the right way to do this
is to fix the server to support multiple clients, not to start up multiple
single-client servers).