Allow forced reload for node types that are not rebootable.
Sometimes we want to mark some node types as not rebootable by default,
but we end up needing to power cycle them occasionally.  For instance, if
we're going to be flashing switches, we don't want any other Emulab
service power cycling them if they are in some unknown state -- only the
service flashing them should be able to power cycle them, because only it
knows what state the node is in.

Tracking node operational state across daemons (or just across fork
boundaries) is a more complex project, and it's not clear it's really
needed, aside from special cases like flashing switches.

So, we just do a little end-around the rebootable attribute.