I've prototyped a change to how tunnels are done. Rather then having to supply the fully specified rspec to both CMs, my changes permit you to specify only the component_manager_urn for the remote side.
When you start the sliver, the CMs contact each other to resolve the
slice (get the manifest), find the actual node in the manifest, and
resolve that node to get the control network.

This even allows for fully unbound rspecs to be supplied, as long as
the component_manager_urn is specified.