Add special case for bbg, same as we do for spp. This stuff needs to be generalized in some fashion.
Change to how vlan ecapsulation links are represented in the lans
table. Formerly, we had two links; one was the underlying vlan and the
other was the ecapsulated link running over it. This is required when
*not* using vlan encapsulation since there might be multiple emulated
networks on top of the vlan network.

However, when using vlan ecapsulation, this strictly one to one; there
are multiple vlan links, with the ports in trunk mode. There is no
reason to have two entries in the lans table, and in fact was causing
me grief wrt vlan tag reservations in protogeni, since the name of
link is not the same as the name of the underlying vlan link that was
getting created. And so it was impossible to match up the
reservations.

I pray this does not break 10 other things.