Frisbee master server compatibility mode fix for vnodes.
The mserver wants to validate nodes by looking up their IP address in the
interfaces table, and vnodes have no interfaces table entries.  So when we
invoke the frisbeehelper in compat mode, do so with the pnode identity
instead. For validation purposes, this is identical.

Is this an issue for non-compat mode? Not right now since all our vnode
implementations that load images make their loadinfo request from the
physical host.  If vnodes start to call the master server on boss directly,
we will have an issue.