(Hopefully) temporary hack to detect null chars in node_ids in the nodes table.
This is completely reactive and based on no reasoning. A common case of the
somebodys-been-messin-with-my-memory bug, is that we start seeing node_ids
with null chars at the end. Here we will send a stack trace to testbed-ops
if this happens (and we strip off the null). I am hoping to narrow down
where this happens.