Assorted changes to make sure that the uid we grab from the user URN is a valid Emulab user id (as for creating accounts on nodes) and for inserting into the Emulab DB.
If the uid is not valid for us, make up a new one from a hash of the
certificate. This will give us a (typically) unique but always
consistent uid to use.

Also add the uid to the services/login section of the manifest so that
the client always knows what uid to use when logging in.