Add dynamic, per-experiment blobs.  Add tables for client side service blobs.
Per-experiment blobs can be created by the parser, and are inserted into
the blobs table at swapin based on what's in virt_blobs.  We go to some
effort to keep any existing filename<->uuid mapping in teh blobs table
so that any caching already done by the server and client in a previous
download of a blob remains intact.  We also remove blobs on experiment
modify that are no longer necessary.  Finally, we remove dynamic blobs
at swapout.