Bug fix: ensure we run any spawned frisbeed as "the appropriate user."
Previously, frisbeed's were always running as the uid/gid of the master server
which generally meant root. This was not a security issue, as we did not rely
on file permissions to enforce accessibility.

For the Emulab config, "appropriate user" means the uid of the experiment
swapper and the gid of the experiment. For the null (default) config, it
still means the uid/gid of the master server--if you don't want them running
as root, don't run the master server as root.

For now, and frisbee clients spawned (e.g., to fetch an image from a parent)
are still run with the uid/gid of the master server. This can easily be
changed if necessary.