Remove the hardwired configuration of physical interfaces used for linkdelays.
I believe this is left over from a bygone era and should not be needed
(in addition to being wrong, since it always sets the link to 100Mb).
rc.ifconfig (via the generated rc.ifc) will properly configure physical
interfaces that underly multiplexed links.

Note also that the Linux version of the delaysetup script never did this.