Hopefully short-term fix to allow more general mixing of pnodes and vnodes.
The current state of affairs is that you can only set the "link emulation"
style at the experiment level, and that forces link emulation on physical
nodes as well as virtual nodes. Thus you are forced into using either 802.11
tagged vlan emulation or an OS that supports something called "veths" on
physical nodes.

From the comment:

   This is a very, very, very special case. If a non-encapsulating veth
   interface (veth-ne) maps 1-to-1 with an underlying physical interface,
   we want to just use the physical interface instead. This allows OSes
   (on physical nodes) which don't support a veth device (i.e., most of
   them) to talk to vnodes which are using veth-ne style.

   This can go away once we have separated the notion of multiplexing
   links from encapsulating links (a historical conflation) so that we
   don't have to force virtual devices onto physical nodes just because
   some virtual nodes in the same experiment require multiplexed links.